Added more easing functions and updated timelapse example
The easing function (used in Animations) was not well-implemented. Implemented a few good-looking easing functions and updated the viewer code to use a CUBIC_OUT animation for the standard rotation (when using the arrow keys). Also updated the TimelapseAnimation example to showcase all easing functions in sequence. Added a "Paused" way to create a timelapse animation, so you can now specifically start the timelapse using Viewer.ResumeTimeLapse()